Move tests to subfolders.
